requirements.txt
selenium~=4.14.0
pytest~=7.4.2
PyYAML~=6.0.1
requests~=2.31.0
pydantic~=2.5.0